requirements.txt
PyYAML
colorama
SQLAlchemy
mitmproxy